Week_1_text-analysis-tool/requirements.txt
openai
google-generativeai
mistralai
transformers
sentencepiece
streamlit
gradio
python-dotenv
transformers
requests
streamlit
python-dotenv
pandas
textblob
langdetect
